refactor : convenience, dto 위치 조정
