+++ src/main/java/com/opus/opus/modules/contest/application/ContestAwardCommandService.java
@@ -3,11 +3,11 @@ package com.opus.opus.modules.contest.application;
 import static com.opus.opus.modules.contest.exception.ContestAwardExceptionType.DUPLICATE_CONTEST_AWARD_NAME;
 import static com.opus.opus.modules.contest.exception.ContestAwardExceptionType.NOT_FOUND_CONTEST_AWARD;
 
-import com.opus.opus.modules.contest.convenience.ContestAwardConvenience;
-import com.opus.opus.modules.contest.convenience.ContestConvenience;
+import com.opus.opus.modules.contest.application.convenience.ContestAwardConvenience;
+import com.opus.opus.modules.contest.application.convenience.ContestConvenience;
 import com.opus.opus.modules.contest.domain.Contest;
 import com.opus.opus.modules.contest.domain.ContestAward;
-import com.opus.opus.modules.contest.dto.request.ContestAwardRequest;
+import com.opus.opus.modules.contest.application.dto.request.ContestAwardRequest;
 import com.opus.opus.modules.contest.exception.ContestAwardException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,7 +22,7 @@ public class ContestAwardCommandService {
     private final ContestAwardConvenience contestAwardConvenience;
 
     public void createContestAward(final Long contestId, final ContestAwardRequest request) {
-        final Contest contest = contestConvenience.getContestById(contestId);
+        final Contest contest = contestConvenience.getValidateExistContest(contestId);
 
         if (contestAwardConvenience.isDuplicateAwardName(contestId, request.awardName())) {
             throw new ContestAwardException(DUPLICATE_CONTEST_AWARD_NAME);
@@ -38,7 +38,7 @@ public class ContestAwardCommandService {
     }
 
     public void updateContestAward(final Long contestId, final Long awardId, final ContestAwardRequest request) {
-        contestConvenience.getContestById(contestId);
+        contestConvenience.getValidateExistContest(contestId);
 
         final ContestAward contestAward = contestAwardConvenience.getContestAwardById(awardId);
 
@@ -54,7 +54,7 @@ public class ContestAwardCommandService {
     }
 
     public void deleteContestAward(final Long contestId, final Long awardId) {
-        contestConvenience.getContestById(contestId);
+        contestConvenience.getValidateExistContest(contestId);
 
         final ContestAward contestAward = contestAwardConvenience.getContestAwardById(awardId);
         if (!contestAward.getContest().getId().equals(contestId)) {
